REFACTOR-#0000: remove code duplication for 'to_*' functions
